volume scheduling: move metrics code into a separate pkg
