+++ pkg/controller/volume/scheduling/metrics/BUILD
@@ -1,16 +1,11 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "metric_recorder.go",
-        "metrics.go",
-    ],
-    importpath = "k8s.io/kubernetes/pkg/scheduler/metrics",
+    srcs = ["metrics.go"],
+    importpath = "k8s.io/kubernetes/pkg/controller/volume/scheduling/metrics",
+    visibility = ["//visibility:public"],
     deps = [
-        "//pkg/controller/volume/scheduling/metrics:go_default_library",
         "//staging/src/k8s.io/component-base/metrics:go_default_library",
         "//staging/src/k8s.io/component-base/metrics/legacyregistry:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
@@ -28,10 +23,5 @@ filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["metric_recorder_test.go"],
-    embed = [":go_default_library"],
+    visibility = ["//visibility:public"],
 )
