+++ pkg/controller/volume/scheduling/metrics/metrics.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scheduling
+package metrics
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
