+++ pkg/controller/volume/scheduling/scheduler_binder_cache.go
@@ -19,7 +19,8 @@ package scheduling
 import (
 	"sync"
 
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/kubernetes/pkg/controller/volume/scheduling/metrics"
 )
 
 // PodBindingCache stores PV binding decisions per pod per node.
@@ -93,7 +94,7 @@ func (c *podBindingCache) DeleteBindings(pod *v1.Pod) {
 
 	if _, ok := c.bindingDecisions[podName]; ok {
 		delete(c.bindingDecisions, podName)
-		VolumeBindingRequestSchedulerBinderCache.WithLabelValues("delete").Inc()
+		metrics.VolumeBindingRequestSchedulerBinderCache.WithLabelValues("delete").Inc()
 	}
 }
 
@@ -113,7 +114,7 @@ func (c *podBindingCache) UpdateBindings(pod *v1.Pod, node string, bindings []*b
 			bindings:      bindings,
 			provisionings: pvcs,
 		}
-		VolumeBindingRequestSchedulerBinderCache.WithLabelValues("add").Inc()
+		metrics.VolumeBindingRequestSchedulerBinderCache.WithLabelValues("add").Inc()
 	} else {
 		decision.bindings = bindings
 		decision.provisionings = pvcs
